backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
requests==2.31.0
pinecone-client==3.0.0
sentence-transformers==2.3.1
torch==2.0.0
python-dotenv==1.0.0
huggingface-hub==0.19.4
transformers==4.35.2
